requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.117.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
aiohttp==3.9.5
nest_asyncio==1.6.0
python-dotenv==1.0.1
